// lib/enums/insuranceTypeEnums.ts
export enum TypeInsurance {
  SCTR = 'Sctr',
  SOAT = 'Soat',
  VIDA_LEY = 'VidaLey',
  POLIZA_CAR = 'PolizaCar',
  POLIZA_TREC = 'PolizaTrec',
  POLIZA_EAR = 'PolizaEar',
  POLIZA_MULTIRRIESGO = 'PolizaMulti',
  R_CIVIL = 'ResCivil',
  VEHICULAR = 'Vehicular',
  VIDA = 'Vida',
  EPS = 'Eps',
  SALUD = 'Salud',
  ACCIDENTES = 'AccidentesP',
  FOLA = 'Fola',
  VIAJES = 'Viajes',
  P_EMPRESARIAL = 'PEmpresarial',
}